evviva che bello

+++ phase2/p2test.c
@@ -115,7 +115,7 @@ void print(char *msg) {
   SYSCALL(PASSEREN, (int)&sem_term_mut, 0, 0); /* P(sem_term_mut) */
   while (*s != EOS) {
     devregtr value = PRINTCHR | (((devregtr)*s) << 8);
-    status = SYSCALL(DOIO, (unsigned int)command, value, 0);
+    status = SYSCALL(DOIO, (int)command, (int)value, 0);
     if ((status & TERMSTATMASK) != RECVD) {
       PANIC();
     }
